+++ app/Http/Controllers/RelicController.php
@@ -27,7 +27,7 @@ class RelicController extends Controller
                     
                 } 
                 if($this->validNumber(request('quantity'))){
-                    if(request('quantity') > 9999){
+                    if((int)request('quantity') > 9999){
                         $item->quantity = 9999;
                     } 
                     else if((int)request('quantity') < 1)
